Fix problem with missing libs/sync/test target.
The problem can be reproduced in a clean regression
test directory by first running tests with --tag=master
and then running tests with --tag=develop.  The fix,
as Beman pointed out, is to init the submodules.
Fortunately the init command is a noop for existing
submodules so added this command to the git testing
sequence when the boost_root directory already exists.

With this patch applied, I am now able to run tests
with master first, and develop second, no error. Tested
on Linux and Darwin.