Display specific failures on summary page.
The failure that's displayed for a library and toolset is the most
significant among all of the failures detected for the toolset. The order
of failures according to which the specific type is choosen is (from the
most significant):
comp > link > run > fail > cerr > time > file

Order the cells in legends according to the failure significance. This
is consistent with the older ordering of the legend elements. They are
sorted from pass to regression failure.